requirements.txt
Django==1.6.5
django-debug-toolbar==1.2.1
pytest-django==2.6.2
selenium==2.42.1
django-widget-tweaks==1.3
factory_boy==2.4.1
django-coffeescript==0.7.2
